% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDRNmir.R
\docType{data}
\name{EDRNmir}
\alias{EDRNmir}
\title{Brain/Liver/Placenta (BLP) miRNA-sequencing mixture data}
\format{Data frame of 16 variables and 968 rows:
\describe{
\item{V1}{mirBase21_hsa identifier}
\item{Mix1_1}{counts from mix 1 ; designed proportions 0.25,0.25,0.5 BLP}
\item{Placenta_1}{Pure Placenta, 1st replicate}
\item{Liver_1}{Pure Liver, 1st replicate}
\item{Mix2_1}{mix 2 ; designed proportions .5,0.25,0.25 BLP}
\item{Placenta_2}{Pure Placenta, 2nd replicate}
}}
\source{
{EDRN Interlab study} {manuscript in preparation}
}
\usage{
EDRNmir
}
\description{
A dataset containing a set of mixes and raw components.
}
\keyword{datasets}

