% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmetrics.R
\name{mixmetrics}
\alias{mixmetrics}
\title{Calculate the fraction of mixture components in a mixture dataset.}
\usage{
mixmetrics(mrnatype = "internalconsensus", splittype = "none",
  prenormalized = TRUE, modeltype = "twomix", indf, retdf = FALSE,
  idnames = "id", idvars = 1, trueproportions, componentnames = c("Brain",
  "Liver", "Placenta"), mixnames = c("Mix1", "Mix2"),
  annot = c("Replicate"), ...)
}
\arguments{
\item{mrnatype}{Internalconsensus,externalagreement,none,or ercc;
How to calculate the measured fraction in a sample.}

\item{splittype}{'none' - not currently implemented:
How to split the data rows based on columns in idvars.}

\item{prenormalized}{TRUE/FALSE - if data is already normalized.}

\item{modeltype}{"twomix","threemix":
How many different mixtures are present in your dataset.}

\item{indf}{Input dataframe:  counts and idvariables.}

\item{idvars}{List of columns (or names?) that are not data, but identifiers
of gene names(1st) or splitting features (not 1st, not implemented).}

\item{trueproportions}{Data frame consisting of the true mixture
proportions of each mix, in order.}

\item{componentnames}{List of column names of indf for each component.}

\item{mixnames}{List of column names of indf corresponding to each mix.}

\item{annot}{How to refer to the various copies of each component/mix
after they are split/merged.}

\item{...}{Anything else.}
}
\description{
\code{mixmetrics} solves a linear model based on the components
and proportions listed.  Returns a modified object with model results and
predicted count values.
}

