% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturesolutions.R
\docType{package}
\name{mixturesolutions-package}
\alias{mixturesolutions}
\alias{mixturesolutions-package}
\title{Solve the proportions of count tables from mixed samples}
\description{
Mixtures allow for known truth / accepted reference values to be defined for a dataset.
Comparing the observed values to these reference values allows you to evaluate performance.
}
\details{
The only function you're likely to need from \pkg{mixpackage} is
\code{\link{mixmetrics}}.
}

