\name{mixturesolutions-package}
\alias{mixturesolutions-package}
\alias{mixturesolutions}
\docType{package}
\title{
\packageTitle{mixturesolutions}
}
\description{
\packageDescription{mixturesolutions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mixturesolutions}
\packageIndices{mixturesolutions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mixturesolutions}

Maintainer: \packageMaintainer{mixturesolutions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
